+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/zz_generated_backupjobs_client.go
@@ -107,7 +107,7 @@ func (client *BackupJobsClient) listCreateRequest(ctx context.Context, vaultName
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-02-01")
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}