+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/zz_generated_backupoperationstatuses_client.go
@@ -102,7 +102,7 @@ func (client *BackupOperationStatusesClient) getCreateRequest(ctx context.Contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header.Set("Accept", "application/json")
 	return req, nil