+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/zz_generated_backuppolicies_client.go
@@ -108,7 +108,7 @@ func (client *BackupPoliciesClient) listCreateRequest(ctx context.Context, vault
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-02-01")
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}