+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/zz_generated_backupworkloaditems_client.go
@@ -119,7 +119,7 @@ func (client *BackupWorkloadItemsClient) listCreateRequest(ctx context.Context, 
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-02-01")
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}