+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/zz_generated_jobs_client.go
@@ -94,7 +94,7 @@ func (client *JobsClient) exportCreateRequest(ctx context.Context, vaultName str
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-02-01")
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}