+++ sdk/resourcemanager/recoveryservices/armrecoveryservicesbackup/zz_generated_protectioncontaineroperationresults_client.go
@@ -110,7 +110,7 @@ func (client *ProtectionContainerOperationResultsClient) getCreateRequest(ctx co
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2021-12-01")
+	reqQP.Set("api-version", "2022-02-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header.Set("Accept", "application/json")
 	return req, nil